# fix(mito): avoid caching empty batches in row group