--- a/src/mito2/src/read/last_row.rs
+++ b/src/mito2/src/read/last_row.rs
@@ -17,7 +17,6 @@
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use common_telemetry::info;
 use store_api::storage::TimeSeriesRowSelector;
 
 use crate::cache::{
@@ -214,7 +213,7 @@ impl RowGroupLastRowReader {
     /// Updates row group's last row cache if cache manager is present.
     fn maybe_update_cache(&mut self) {
         if let Some(cache) = &self.cache_manager {
-            if self.yielded_batches.len() == 0 {
+            if self.yielded_batches.is_empty() {
                 // we always expect that row groups yields batches.
                 return;
             }