--- a/tests/cases/standalone/common/select/last_value.result
+++ b/tests/cases/standalone/common/select/last_value.result
@@ -1,0 +1,69 @@
+create table t (
+    ts timestamp time index,
+    host string primary key,
+    not_pk string,
+    val double,
+) with (append_mode='true');
+
+Affected Rows: 0
+
+insert into t values
+    (0, 'a', '🌕', 1.0),
+    (1, 'b', '🌖', 2.0),
+    (2, 'a', '🌗', 3.0),
+    (3, 'c', '🌘', 4.0),
+    (4, 'a', '🌑', 5.0),
+    (5, 'b', '🌒', 6.0),
+    (6, 'a', '🌓', 7.0),
+    (7, 'c', '🌔', 8.0),
+    (8, 'd', '🌕', 9.0);
+
+Affected Rows: 9
+
+select flush_table('t');
+
++------------------------+
+| flush_table(Utf8("t")) |
++------------------------+
+| 0                      |
++------------------------+
+
+select
+        last_value(host order by ts),
+        last_value(not_pk order by ts),
+        last_value(val order by ts)
+from t
+    group by host
+    order by host;
+
++---------------------------------------------------+-----------------------------------------------------+--------------------------------------------------+
+| last_value(t.host) ORDER BY [t.ts ASC NULLS LAST] | last_value(t.not_pk) ORDER BY [t.ts ASC NULLS LAST] | last_value(t.val) ORDER BY [t.ts ASC NULLS LAST] |
++---------------------------------------------------+-----------------------------------------------------+--------------------------------------------------+
+| a                                                 | 🌓                                                  | 7.0                                              |
+| b                                                 | 🌒                                                  | 6.0                                              |
+| c                                                 | 🌔                                                  | 8.0                                              |
+| d                                                 | 🌕                                                  | 9.0                                              |
++---------------------------------------------------+-----------------------------------------------------+--------------------------------------------------+
+
+-- repeat the query again, ref: https://github.com/GreptimeTeam/greptimedb/issues/4650
+select
+        last_value(host order by ts),
+        last_value(not_pk order by ts),
+        last_value(val order by ts)
+from t
+    group by host
+    order by host;
+
++---------------------------------------------------+-----------------------------------------------------+--------------------------------------------------+
+| last_value(t.host) ORDER BY [t.ts ASC NULLS LAST] | last_value(t.not_pk) ORDER BY [t.ts ASC NULLS LAST] | last_value(t.val) ORDER BY [t.ts ASC NULLS LAST] |
++---------------------------------------------------+-----------------------------------------------------+--------------------------------------------------+
+| a                                                 | 🌓                                                  | 7.0                                              |
+| b                                                 | 🌒                                                  | 6.0                                              |
+| c                                                 | 🌔                                                  | 8.0                                              |
+| d                                                 | 🌕                                                  | 9.0                                              |
++---------------------------------------------------+-----------------------------------------------------+--------------------------------------------------+
+
+drop table t;
+
+Affected Rows: 0
+