I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)

- fixes #4650

## Performance regression

Run the same TSBS query with `region_engine.mito.selector_result_cache_size=2GB`:

Before:
```
$ ./bin/tsbs_run_queries_influx --file=./bench-data/greptime-queries-lastpoint.dat     --db-name=benchmark     --urls="http://localhost:4000"
Run complete after 10 queries with 1 workers (Overall query rate 2.61 queries/sec):
Influx last row per host:
min:    14.68ms, med:    28.02ms, mean:   383.09ms, max: 3610.37ms, stddev:  1075.75ms, sum:   3.8sec, count: 10
all queries             :
min:    14.68ms, med:    28.02ms, mean:   383.09ms, max: 3610.37ms, stddev:  1075.75ms, sum:   3.8sec, count: 10
wall clock time: 3.870371sec
```

After:

```
$ ./bin/tsbs_run_queries_influx --file=./bench-data/greptime-queries-lastpoint.dat     --db-name=benchmark     --urls="http://localhost:4000"
Run complete after 10 queries with 1 workers (Overall query rate 2.89 queries/sec):
Influx last row per host:
min:   307.68ms, med:   334.83ms, mean:   345.94ms, max:  425.07ms, stddev:    31.58ms, sum:   3.5sec, count: 10
all queries             :
min:   307.68ms, med:   334.83ms, mean:   345.94ms, max:  425.07ms, stddev:    31.58ms, sum:   3.5sec, count: 10
wall clock time: 3.473471sec
```

Within noise threshold.


## Checklist

- [x] I have written the necessary rustdoc comments.
- [x] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.